Product CRUD Processes

+++ app/Http/Controllers/admin/AdminProductController.php
@@ -4,27 +4,12 @@ namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class AdminProductController extends Controller
 {
-
-    protected $appends=[
-        'getParentsTree'
-    ];
-
-    public static function getParentsTree($product,$title)
-    {
-        if($product->parent_id == 0){
-            return $title;
-        }
-        $parent = Product::find($product->parent_id);
-        $title=$parent->title .'>'.$title;
-        return CategoryController::getParentsTree($parent,$title);
-
-
-    }
 
 
 
@@ -50,7 +35,7 @@ class AdminProductController extends Controller
      */
     public function create()
     {
-        $data=Product::all();
+        $data=Category::all();
         return view('admin.product.create',[
             'data'=>$data
         ]);
@@ -66,10 +51,13 @@ class AdminProductController extends Controller
     public function store(Request $request)
     {
         $data = new Product();
-        $data->parent_id=$request->parent_id;
-        $data->title=$request->title;
-        $data->keywords=$request->keywords;
-        $data->description=$request->description;
+        $data->category_id = $request->category_id;
+        $data->user_id = 0; //$request->category_id;
+        $data->title = $request->title;
+        $data->keywords = $request->keywords;
+        $data->description = $request->description;
+        $data->detail = $request->detail;
+        $data->city = $request->city;
         $data->status=$request->status;
         if ($request->file('image')){
             $data->image=$request->file('image')->store('images');
@@ -81,7 +69,7 @@ class AdminProductController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function show(Product $product, $id)
@@ -99,11 +87,11 @@ class AdminProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function edit(Category $product,$id)
+    public function edit(Product $product,$id)
     {
 
         $data= Product::find($id);
-        $datalist=Product::all();
+        $datalist=Category::all();
         return view('admin.product.edit',[
             'data'=>$data,
             'datalist'=>$datalist
@@ -120,16 +108,24 @@ class AdminProductController extends Controller
     public function update(Request $request, Product $product, $id)
     {
         $data = Product:: find($id);
-        $data->parent_id = $request->parent_id;
+        $data->category_id = $request->category_id;
+        $data->user_id = 0; //$request->category_id;
         $data->title = $request->title;
         $data->keywords = $request->keywords;
         $data->description = $request->description;
+        $data->detail = $request->detail;
+        $data->type = $request->type;
+        $data->city = $request->city;
+        $data->country = $request->country;
+        $data->konum = $request->konum;
         $data->status = $request->status;
         if ($request->file('image')){
             $data->image=$request->file('image')->store('image');
         }
         $data->save();
         return redirect('admin/product');
+
+
 
     }
         /**
@@ -143,7 +139,9 @@ class AdminProductController extends Controller
         {
             //
             $data= Product::find($id);
-            Storage::delete($data->image);
+            if($data->image && Storage::disk('public')->exists($data->image)){
+                Storage::delete($data->image);
+            }
             $data->delete();
             return redirect('admin/product');
         }
